frontend

+++ frontend/src/app/contacts/contacts-list/contacts-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, ChangeDetectorRef, OnDestroy } from '@angular/core';
 import { Observable, Subscription } from 'rxjs';
-import { ContactService, UsersService } from '../../shared/api-generated/api-generated';
+import { ContactService, UsersService, DataProtectionService } from '../../shared/api-generated/api-generated';
 import { ContactDto } from '../../shared/api-generated/api-generated';
 import { MatDialog } from '@angular/material/dialog';
 import { ContactsAddHistoryComponent } from '../contacts-add-history/contacts-add-history.component';
@@ -10,6 +10,8 @@ import { ContactsEditDialogComponent } from '../contacts-edit-dialog/contacts-ed
 import { ContactsAddDialogComponent } from '../contacts-add-dialog/contacts-add-dialog.component';
 import { MediaObserver, MediaChange } from '@angular/flex-layout';
 import { JwtService } from 'src/app/shared/jwt.service';
+import { DpUpdatePopupComponent } from 'src/app/shared/data-protection/dp-update-popup/dp-update-popup.component';
+import { DataProtectionHelperService } from 'src/app/shared/data-protection';
 
 @Component({
   selector: 'app-contacts-list',
@@ -31,6 +33,8 @@ export class ContactsListComponent implements OnInit, OnDestroy {
     private changeDetectorRefs: ChangeDetectorRef,
     private dialog: MatDialog,
     private mediaObserver: MediaObserver,
+    private readonly dataProtectionService: DataProtectionService,
+		  private readonly dsgvoService: DataProtectionHelperService,
     private jwt: JwtService) {
       this.flexMediaWatcher = mediaObserver.asObservable().subscribe((change: MediaChange[]) => {
       if (change[0].mqAlias !== this.currentScreenWidth) {
@@ -76,11 +80,22 @@ export class ContactsListComponent implements OnInit, OnDestroy {
   openEditDialog(contact: ContactDto) {
     const dialogRef = this.dialog.open(ContactsEditDialogComponent, { data: contact, disableClose: true });
 
-    dialogRef.afterClosed().subscribe((result) => {
-      if (result.delete) {
-        this.deleteContact(result.id);
+    dialogRef.afterClosed().subscribe((editDialogResult) => {
+      if (editDialogResult.delete) {
+        this.deleteContact(editDialogResult.id);
+      } else {
+        if (editDialogResult.newContact && editDialogResult.newContact) {
+        const dialogDSGVORef = this.dialog.open(DpUpdatePopupComponent, {disableClose: true});
+
+        dialogDSGVORef.afterClosed().subscribe(sendMessage => {
+          if (sendMessage) {
+            const diff = this.dsgvoService.getDiffOfObjects( editDialogResult.newContact, editDialogResult.oldContact, ['unchanged']);
+            this.dataProtectionService.sendMessage({}).subscribe();
+          }
+        });
       }
-      this.getData();
+        this.getData();
+      }
     });
   }
 
@@ -92,7 +107,15 @@ export class ContactsListComponent implements OnInit, OnDestroy {
 
     deleteDialogRef.afterClosed().subscribe((deleteResult) => {
       if (deleteResult?.delete) {
-        this.service.delete(id).subscribe(x => this.getData());
+
+        const dialogDSGVORef = this.dialog.open(DpUpdatePopupComponent, {disableClose: true});
+
+        dialogDSGVORef.afterClosed().subscribe(sendMessage => {
+          if (sendMessage) {
+            this.dataProtectionService.sendDeleteMessage({}).subscribe();
+            this.service.delete(id).subscribe(x => this.getData());
+          }
+        });
       }
     });
   }
@@ -123,6 +146,6 @@ export class ContactsListComponent implements OnInit, OnDestroy {
         phoneNumber: '0172-9344333' + this.length,
         contactEntries: []
       }
-    }, this.jwt.getUserId()).subscribe(x => this.getData());
+    }).subscribe(x => this.getData());
   }
 }
